+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Quiqr App CHANGELOG
-## **WORK IN PROGRESS**
+## 0.18.7 (2024-04-30)
 - hide dogfood from cms model parseinfo
 - new method to add includes: #500
     - quiqr/model/includes/{singles,collections, menus}/file.yaml